Add configuration from command line via YAML (#17)
1) User can overwrite default configuration via YAML file e.g.
   $> doclassification_regression -c myConfig.yml

2) User can get default config in YAML via
   $> doclassification_regression -d $DEFAULT_CONFIG_YAML
   to path $DEFAULT_CONFIG_YAML

All missing parameters are replaced by default parameters. In case of an
unknown parameter or if a value is missing for a parameter the execution
is aborted. That can make sense if the user made a typo like 'dotaining'
instead of 'dotraining'.

Any parameter - value combination missing is set to default.

Parameter dependencies are checked to first order (such as ROC cannot
be calcualted for mltype == "Regression"). That means if a dependent
parameter has another dependency it will only check param1 -> param2
but not param1 -> param2 -> param3. Furthermore, cyclic dependencies
are not yet detected.