Add logger utility
The logger utility can be used for central logging and all verbosity.
The standard log level is INFO. If a critical message is sent, it will
be flushed and the execution will be aborted afterwards.

Using the logger is highly recommended and provides a unified format for
the verbosity. All normal print commands have hence been replaced by
logger.info

The icentral logger object can be obtained by importing the 'get'
function get_logger via
from machine_learning_hep.logger import get_logger

and then one can do

logger = get_logger()
logger.info("Some message")

On the command line, debug information can be enabled using the --debug
flag such that messages logged with logger.debug are printed as well.

The log can be additionaly written to a file by specifying the
--logfile <logfilepath>
Note that the logs are appended to the logfile if it already exists.