Use only a fraction of real data in significance optimization
Only a fraction of real data is used to get the background for the significance estimation. There isn't any bias related to the application of a model on the same candidates used for the training, if the sum of the candidates employed in the training and in the significance optimization is less than the total available.