Implement selection compatible with bitmap and old method (#79)
* Add function to select dataframe

A new function to apply different candidate selections to a daframe is added, the parameters are currently implemented only for Ds

* Use new function in signif. opt.

* Remove .pkl dataframes when cleaning

* Propagate new signal selection to other modules

* Update Ds database

* Add debug messages to filter_df_cand

* Remove unused variables

* Update other particle database

* Fix for webapp

* Remove variables not yet available in data trees

cand_fileID_ML and cand_evtID_ML are temporary removed because they are not present in the current version of data.
With them some functions crash, e.g. do_correlation(...). To be added again later.

* Add back var_others

* Remove old test file

* Update comments