Temporary fix for logging
The newest version of abseil (https://github.com/abseil/abseil-py)
apparently interferes with user logging implemented via

import logging

For details see https://github.com/abseil/abseil-py/issues/99

This was observed when installing the ML package from scratch during the
last week. Comparing TensorFlow versions it was observed that most
people's ML package was still using TF version 1.13.1/2. Installing the
package now leads to installing TF vesion 1.14 and hence a more recent
version of abseil as well which causes the weird logging behaviour.

A fix is possible thanks to
https://github.com/abseil/abseil-py/issues/99#issuecomment-513414324

As soon as this is fixed in abseil this will be propagated to the ML
package.