Refine fitting (#435)
* MC init fit to single gaussian

* Refine initila MC fit

The initial MC fit in the mult. int. bin is now done on pure MC signal
invariant mass histogram.
This fit is repeated with an increasing number of RMS ranges around the
initially guessed mean of the invariant mass distribution.
As soon as the fitted sigma does not exceed the RMS by more than 10%,
this is taken to be the sigma from the initial MC fit.
This procedure is done to keep the influence of the (potentially) large
tails present in the distribution which would bias the sigma towards
larger values.

Furthermore, if the fit with the initialiation procedure chosen by the
user fails, this is clearly marked in the respective final fit canvas.

* Take mult. int. rebinning for init data fit