Make (multi)analyzer analysis calls flexible
The multianalyzer calls the analyzer base class with the desired
analysis steps as string. The presence of a corresponding method is
checked which is then executed. Non-present analysis steps are collected
and reported to be missing at the end.
In that way the multianalyzer does not need to be aware of any analysis
method but just tries the analyzer classes.