Revive grid search
Multiple scorings can be chosen to be extracted during grid search. At
the moment "AUC" and "Accuracy" are available. One of that has to be
chosen for the "refit" option and based on that the best classifier will
be found.

Summarize means and stds of train and test scores per model and per
training bin providing one sub-plot per score. The score chosen for
"refit" is always the one on top. Indices of used parameter are on the
x-axis and the lists of parameters are written in the plot to easily
identify the best model.

Moving grid search parameters from database_ml_gridsearch.yml to
config_model_parameters.yml to have everything in one place.