Implement the lazy mode.
Avoid running the processor for variations when it is not needed and use the default processor output as the analyzer input.
Each variation category has an attribute "processor" to specify whether running the processor with the modified database is required.
If keys "data_proc" and "mc_proc" exist in the analysis database, the "results" paths there will be used as the analyzer input.
If a variation requires the processor to run, the analysis will run with the "default_complete.yml" configuration file, otherwise the file "default_analyzer.yml" will be used to run only the analyzer part of the analysis.
The optional parameter -p allows to process either only variations that need the processor (value 1) or only those that do not (value 0).